LIG-9843: split evaluation route tests per function
Replace the monolithic test_evaluation.py with a tests/.../evaluation/ package:
shared helpers + one test module per route (get_runs, get_sample_metrics_info,
get_confusion_matrix).
